// packages/backend/src/utils/enums/EvaluationsEnum.ts
export enum EvaluationsEnum {
	CAP = 'cap',
	EVALUATION_ID = 'evaluation_id',
	MULTI_FAMILY = 'multi_family',
	LEASE = 'lease',
	SALE = 'sale',
	COST = 'cost',
	INCOME = 'income',
	LAND_TYPE = 'land_type',
	MOST_LIKELY_OWNER_USER = 'most_likely_owner_user',
	TOPOGRAPHY = 'topography',
	LOCATION = 'location',
	LOT_SHAPE = 'lot_shape',
	FRONTAGE = 'frontage',
	UTILITIES_SELECT = 'utilities_select',
	CONDITION = 'condition',
	PARKING = 'parking',
	DATE_SOLD = 'date_sold',
	TYPE = 'type',
	COMP_TYPE = 'comp_type',
	COMPARISON_BASIS = 'comparison_basis',
	BASEMENT = 'basement',
	EXTERIOR = 'exterior',
	ROOF = 'roof',
	ELECTRICAL = 'electrical',
	PLUMBING = 'plumbing',
	HEATING_COOLING = 'heating_cooling',
	WINDOWS = 'windows',
	ADA_COMPLIANCE = 'ada_compliance',
	PROPERTY_CLASS = 'property_class',
	PROPERTY_RIGHTS = 'property_rights',
	MAIN_STRUCTURE_BASE = 'main_structure_base',
	FOUNDATION = 'foundation',
	LAND_ONLY = 'land_only',
	SF = 'sf',
	UNIT = ' unit',
	BED = ' bed',
	BUILDING_WITH_LAND = 'building_with_land',
	TYPE_MY_OWN = 'Type My Own',
	ID = 'id',
	RENT_ROLL = 'rent_roll',
	EVALUATION_INCOME_APPROACH_ID = 'evaluation_income_approach_id',
	ROADMAP = 'roadmap',
	PRICE_SF = '$/SF',
	PRICE_ACRE = '$/Acre',
	ORDER_BY_COLUMN = 'order',
	COST_AREA_MAP = 'save-cost-area-map',
	SALE_AREA_MAP = 'save-sale-area-map',
	LEASE_AREA_MAP = 'save-lease-area-map',
	CAP_AREA_MAP = 'save-cap-area-map',
	MULTI_FAMILY_AREA_MAP = 'save-multi-family-area-map',
	HAS_INCOME_APPROACH = 'has_income_approach',
	HAS_SALE_APPROACH = 'has_sales_approach',
	HAS_COST_APPROACH = 'has_cost_approach',
	ACRE = 'ACRE',
	LAND_ONLY_TYPE = 'Land Only',
	SQ_FT = 'sq_ft',
	COMPARISON_BASIS_SF = 'SF',
	COMPARISON_BASIS_UNIT = 'UNIT',
	COMPARISON_BASIS_BED = 'BED',
	SF_SOURCE = 'sf_source',
}

export enum IncomeApproachEnum {
	APPRAISAL_INCOME_APPROACH_ID = 'appraisal_income_approach_id',
	RENT_SQ_FT = 'rent_sq_ft',
	RENT = 'rent_',
	TOTAL_RENT = 'total_rent',
	HIGH_CAP_RATE = 'high_capitalization_rate',
	TOTAL_OE_PER_SQUARE_FEET = 'total_oe_per_square_feet',
	TOTAL_PER = 'total_per',
	TOTAL = 'total',
	TOTAL_OE_PER = 'total_oe_per',
	CAPITALIZATION_RATE = 'capitalization_rate',
	SQ_FT_CAP_RATE = 'sq_ft_capitalization_rate',
	INDICATED_RANGE = 'indicated_range',
	INDICATED_RANGE_SQ_FT = 'indicated_range_sq_ft',
}